requirements/requirements.txt
# Python standard library imports
python-dateutil==2.9.0.post0
requests>=2.32.0
httpx==0.27.2

# third party imports
## django and misc
dj_database_url==2.1.0
django
django-filter==23.5
django-htmx==1.17.3
django-two-factor-auth==1.16.0
django-simple-captcha==0.6.0
django-citext==1.0.0
docutils==0.20.1
markdown
pandas
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-dotenv==1.0.1
azure-identity==1.17.1
fontawesomefree==6.6.0

# We don't use this but django-two-factor-auth requires it
# They are working to make it optional https://github.com/jazzband/django-two-factor-auth/issues/469
phonenumbers==8.13.31

## graphing
plotly==5.19.0

# NHS number
nhs-number

# live application server
gunicorn>=22.0.0

# code linting and formatting
autopep8==2.0.4
black>=24.3.0

# testing and code analysis
coverage==7.4.3
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-asyncio==0.24.0

# versioning
bump2version

# DOCUMENTATION - MATERIAL FOR MKDOCS INSIDERS
# IMPORTANT: This project uses Material for MkDocs **INSIDERS** Edition.
# To install this you will need a GitHub token which is available (for RCPCH team only)
# in our .env files

# git+https://${MATERIAL_FOR_MKDOCS_INSIDERS_TOKEN}@github.com/squidfunk/mkdocs-material-insiders.git

# Further information is at https://squidfunk.github.io/mkdocs-material/insiders/getting-started/

mkdocs-material==9.5.13
mkdocs-git-committers-plugin-2==2.3.0 # displays authors at the bottom of the page
mkdocs-git-revision-date-localized-plugin==1.2.4 # displays last edit date at the bottom of the page
mkdocs-macros-plugin==1.0.5 # enables 'foldable' admonition text (used for large code blocks)
mkdocs-with-pdf==0.9.3 # PDF export feature

# for colored logs
colorlog==6.8.2
